fix: cast statvfs block counts to u64 on macOS

+++ src/disk_space.rs
@@ -195,8 +195,8 @@ fn query_disk_space_at(path: &Path) -> Option<DiskSpace> {
         return None;
     }
     let frsize = stat.f_frsize as u64;
-    let available_bytes = stat.f_bavail.saturating_mul(frsize);
-    let total_bytes = stat.f_blocks.saturating_mul(frsize);
+    let available_bytes = (stat.f_bavail as u64).saturating_mul(frsize);
+    let total_bytes = (stat.f_blocks as u64).saturating_mul(frsize);
     if total_bytes == 0 {
         return None;
     }
